refactor(datasource): extract glob-path splitting into listing_glob module
Move GLOB_START_CHARS and split_glob_expression (with its unit tests) out
of url.rs into a new datafusion/datasource/src/listing_glob.rs module.
No behaviour change. Sets up subsequent commits that grow the glob
helpers (DuckDB-compatible matching options, URL-path glob splitting)
without further bloating url.rs.
